requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=4.5=1_gnu
blas=1.0=mkl
brotli=1.0.9=h9c3ff4c_4
bzip2=1.0.8=h7b6447c_0
ca-certificates=2021.10.8=ha878542_0
certifi=2021.10.8=py38h578d9bd_1
colorama=0.4.4=pyh9f0ad1d_0
cudatoolkit=10.2.89=hfd86e86_1
cycler=0.11.0=pyhd8ed1ab_0
dbus=1.13.18=hb2f20db_0
einops=0.3.2=pyhd8ed1ab_0
expat=2.2.10=h9c3ff4c_0
ffmpeg=4.3=hf484d3e_0
fontconfig=2.13.1=hba837de_1005
fonttools=4.25.0=pyhd3eb1b0_0
freetype=2.11.0=h70c0345_0
gettext=0.19.8.1=hf34092f_1004
giflib=5.2.1=h7b6447c_0
glib=2.58.3=py38h73cb85d_1004
gmp=6.2.1=h2531618_2
gnutls=3.6.15=he1e5248_0
gst-plugins-base=1.14.5=h0935bb2_2
gstreamer=1.14.5=h36ae1b5_2
icu=64.2=he1b5a44_1
intel-openmp=2021.4.0=h06a4308_3561
jpeg=9d=h7f8727e_0
kiwisolver=1.3.1=py38h2531618_0
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
libedit=3.1.20210910=h7f8727e_0
libffi=3.2.1=hf484d3e_1007
libgcc-ng=9.3.0=h5101ec6_17
libgfortran-ng=7.5.0=ha8ba4b0_17
libgfortran4=7.5.0=ha8ba4b0_17
libgomp=9.3.0=h5101ec6_17
libiconv=1.15=h63c8f33_5
libidn2=2.3.2=h7f8727e_0
libpng=1.6.37=hbc83047_0
libstdcxx-ng=9.3.0=hd4cf53a_17
libtasn1=4.16.0=h27cfd23_0
libtiff=4.2.0=h85742a9_0
libunistring=0.9.10=h27cfd23_0
libuuid=2.32.1=h7f98852_1000
libuv=1.40.0=h7b6447c_0
libwebp=1.2.0=h89dd481_0
libwebp-base=1.2.0=h27cfd23_0
libxcb=1.13=h7f98852_1003
libxml2=2.9.10=hee79883_0
lz4-c=1.9.3=h295c915_1
matplotlib=3.4.3=py38h578d9bd_1
matplotlib-base=3.4.3=py38hbbc1b5f_0
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py38h7f8727e_0
mkl_fft=1.3.1=py38hd3c417c_0
mkl_random=1.2.2=py38h51133e4_0
munkres=1.1.4=pyh9f0ad1d_0
ncurses=6.3=h7f8727e_2
nettle=3.7.3=hbbd107a_1
numpy=1.21.2=py38h20f2e39_0
numpy-base=1.21.2=py38h79a1101_0
olefile=0.46=pyhd3eb1b0_0
openh264=2.1.1=h4ff587b_0
openssl=1.1.1l=h7f8727e_0
pcre=8.45=h9c3ff4c_0
pillow=8.4.0=py38h5aabda8_0
pip=21.2.4=py38h06a4308_0
pthread-stubs=0.4=h36c2ea0_1001
pyparsing=3.0.6=pyhd8ed1ab_0
pyqt=5.9.2=py38h05f1152_4
python=3.8.0=h0371630_2
python-dateutil=2.8.2=pyhd8ed1ab_0
python_abi=3.8=2_cp38
pytorch=1.10.1=py3.8_cuda10.2_cudnn7.6.5_0
pytorch-mutex=1.0=cuda
qt=5.9.7=h0c104cb_3
readline=7.0=h7b6447c_5
scipy=1.7.1=py38h292c36d_2
setuptools=58.0.4=py38h06a4308_0
sip=4.19.13=py38he6710b0_0
six=1.16.0=pyhd3eb1b0_0
sqlite=3.33.0=h62c20be_0
timm=0.4.12=pyhd8ed1ab_0
tk=8.6.11=h1ccaba5_0
torchaudio=0.10.1=py38_cu102
torchvision=0.11.2=py38_cu102
tornado=6.1=py38h497a2fe_1
tqdm=4.62.3=pyhd8ed1ab_0
typing_extensions=3.10.0.2=pyh06a4308_0
wheel=0.37.0=pyhd3eb1b0_1
xorg-libxau=1.0.9=h7f98852_0
xorg-libxdmcp=1.1.3=h7f98852_0
xz=5.2.5=h7b6447c_0
zlib=1.2.11=h7f8727e_4
zstd=1.4.9=haebb681_0


